Move org-roam-mode-map key bindings to org-mode-map
I thought org-roam-mode-map is enabled only when editing org roam
files. However, it didn't. org-roam-mode is enabled always. So I moved
key maps to org-mode.

It is not the way I wanted. I wanted that the key bindings are enabled
only when editing org roam files. The current version is best way to
do it with my current knowledge.